requirements.txt
kopf[uvloop]
openstacksdk
